requirements.txt
# Backend
fastapi==0.116.0
uvicorn==0.35.0
python-multipart==0.0.20

# AI & RAG
langchain==0.3.26
langchain-core==0.3.66
langchain-community==0.3.27
langchain-openai==0.3.2
qdrant-client==1.14.2
openai==1.91.0
litellm==1.74.0
fastembed==0.7.1

# Optional: Sentence Transformers (large dependencies ~1.2GB)
# Uncomment if you want to use sentence_transformers provider
# langchain-huggingface==0.3.0
# sentence-transformers==5.0.0
# transformers==4.53.1
# torch==2.7.1

# Database
sqlalchemy==2.0.41
alembic==1.16.2
psycopg2-binary==2.9.9

# Config & Utilities
python-dotenv==1.0.1
pyyaml==6.0.2

# Development & Debugging
debugpy==1.8.14

# Playground
streamlit==1.44.0
requests==2.32.3

# Document Loaders
pypdf==5.4.0
unstructured[md]==0.18.3
langchain-unstructured==0.1.6

# Web Content Extraction
beautifulsoup4==4.12.3
lxml==5.1.0
html2text==2024.2.26